[ #48035 ] Add ExtendedVideoDeviceModel
This new model is a proxy of the original one, but with additional
"special" devices with as still image, screen sharing or sflphone logo

It doesn't work very well with the new daemon validation feature.
This is a design limitation that will be addressed in the daemon.
Until this happen, changing resolution will leave the settings
in an invalid state for some milliseconds, so the client will reset
them just to currupt them again. See issue #48203 for the design
issues.