[ #53076 #53074 ] Store contact information in history log
This commit refactor the contact handling code so it can merge
multiple contacts into one at any point in time. This allow
SFLPhone-KDE to store contacts UID in history, then start to
use them before Akonadi or an other backend start. This is
the first part of a 3 commit change that will also allow
PhoneNumbers to be merged. In turn, this will double the ammount
of data used to match calls to contact, solving bug #52259

This commit also remove some underused constants and add comments