[ #53076 ] Rewrork the phone directory ::getNumber() to take into account more factors
This is quite different than the previous code path. I am not sure this will
cause no regression, time will tell. At least it helps a lot when it come
to getting good PhoneNumber candidate and auto-cleanup the existing ones.

As there is an infinite ammount of corner case, this new code path will
be improved in the future. It is cleaner and easier to work with than the old one.