[ #64375 ] Merge the profile branch from Alex Lision
The current profile model is not bug free, but as there is a lot
of changes made to LibRingClient, I merge that now before it get
even harder to merge. The current profile model is affected by
a race condition with AccountModel where new accounts will be
forgotten. As this is a new feature and it never existed before,
I merge that as a work in progress code.