Rename Contact to Person
The rational for this is that the Contact class was
sometime used for things that has nothing to do with contacts,
such as user profile or network buddy presence. The Contact
object also isn't enough to contact someone, the PhoneNumber
concept (soon to be renamed ContactMethod) is used for this.

Refs #65335