macro: Move the code from libkring to LRC
The KDE client had this code because it had external dependencies.
Those dependencies have been merged into Qt5Core and thus can be merged
into LRC.

The macro system allow keyboard shortcut or other user
experience element to be executed. For example, it allow to attach
a keyboard shortcut to call someone, even if it has a phone "menu"
behind the number.

Refs #73706

Change-Id: Idff34193e229144cba0412ca07fecae03e0abe40