Update patch set 2
Patch Set 2: Code-Review-1

Just FYI

The issue with this solution is that it cause a lot of code to be included by call.h and create multiple copies of that function for nothing. The original solution "almost" work on Windows.

* The type incremented work as expected
* The type ids are correctly generated
* The hash-map is filled, but no correctly
* When a query happen on the missing type, "0" is returned and later converted to Media::Type::AUDIO ( because the default hash value is 0 ).

I would rather fix the old implementation before resorting to merge this one. However, it does provide a working work-around for the time being.

Patch-set: 2
Label: Code-Review=-1