quality: Fix an asan issues with LLVM on Linux
GCC doesn't trigger asan for this, but LLVM seem to refuse
to "not-initialize" the m_lData array and instead force it
to be a pointer to remove the initilization issue. The data
are always initialized in the constructor, but in the ctor
body. Apparently, this is too late for the LLVM version of
asan.

Refs #74376

Change-Id: I15dca26facd8292899d3f36b6b2802ed2f000630