useractionmodel: Port to the 3rd gen Matrix API
It was still using the first C++0x/GCC 4.6 version.

This commit also add new Actions, context and assets to
the UAM. This will allow history, call, bookmarks and
contact to be managed by a single codepath. This should
allow to remove around 1200 line of code from the overall
system which implement context menus for all assets.

Refs #74590

Change-Id: Id2aa8fb21c72493a0c9eeebae13edd7e02b8e949