person: add getter for placeholder attribute
This will help fix "add to contact" client's logic. To display the
add-to-contact form in the clients we need to check if the ContactMethod is
already linked to a Person or not. Problem is, it can be linked to a placeholder
that will never be filled (addressbook not working, person deleted...)

So we need to check if the contactmethod is linked to a placeholder. If it is
we, for now, assume that the Person associated to it can be overriden.

Issue: #78236
Change-Id: Ib84af4cb9e586596e0463846c7bda9d39c3160a3