Update patch set 1
Patch Set 1:

As Stepan said, this is a fist implementation to enable an easy access
to the feature. We need it now: credential is also right and maybe a
better solution, but also a longer to implement solution. Not possible
in our context.
Anyway: we need to fix cland build first...

Patch-set: 1
Label: Code-Review=0