Update patch set 6
Patch Set 6:

I am not sure. Until now, this seem to be very platform dependant. On Linux, all solution usually work. On windows, the C:\ and backslash are required for the name to be parsed (or use file://). You have to check with edric if any solution proposed here work on all platform. Reverting to QString could also mitigate the problem. While QUrl is technically the right class for this, it has caused some issues (like this) in the past

Patch-set: 6
Label: Code-Review=0