interfaces: refactor delegates
*WARNING* this patch breaks the delegate API

The "delegate" design pattern was not really followed by most of the
delegates. They acted more as interfaces used by LRC and occasionally
the client with global instances provided either by LRC or the client.

Thus they have been refactored as such. The "interfaces" folder and
"Interfaces" namespace now contain abstract classes which define the
various interfaces. Their global instances are set and accessed via
the API provided by "globalinstances.h". LRC provides a default
implementation of most of the interfaces and this implementation is
the one used if none other is provided by the client. If no
implementation is set and the global instance is attempted to be
accessed an exception is thrown.

The "globalinstances.h" API allows us to get rid of the singleton
pattern of the delegates, prevents the possibility of the "static
initialization order fiasco", and manages the object lifecycle of
the interfaces instances.

Issue: #79656
Change-Id: I4ee86150a6a92b16402cb72e46838858f49f742e