delegates: refactor delegate management
*WARNING* this patch breaks the delegate API

The lifecycle of delegate objects is now managed by the DelegateManager.
LRC and clients should now get a shared_ptr to the manager via getDelegateManager()
and then get and set the delegate(s) they need via the manager.
If LRC provides a non-abstarct implementation of a given delegate, it will now be
instantiated the first time a get of that delegate is called on the manager. Otherwise
a nullptr will be returned by the get.

This refactoring is done in order to prevent the "static initialization order fiasco"
and to ensure proper delegate object lifecycle management.

Issue: #79348
Change-Id: I4ee86150a6a92b16402cb72e46838858f49f742e