interfaces: refactor delegates
*TODO: update commit msg*

*WARNING* this patch breaks the delegate API

The delegates are now placed in the "Delegates" namespace and their
instances are managed by the global getter and setter functions in
delegates/delegates.h

If LRC provides a non-abstarct implementation of a given delegate, it
will now be instantiated the first time the getter of that delegate is
called. Otherwise a nullptr will be returned by the get.

This refactoring is done in order to prevent the "static initialization
order fiasco" and to ensure proper delegate object lifecycle management.

Issue: #79656
Change-Id: I4ee86150a6a92b16402cb72e46838858f49f742e