Update patch set 7
Patch Set 7:

(1 comment)

> Ideally, yes; however there are instances of the clients using a delegate directly in cases where an lrc model is missing a call to one... so its not strictly true until lrc is patched (and consecutively the clients) for those cases

Ah, the pixel one, you, I forgot about it. The problem was that, unless SVG are used (and they are expensive to work with), LRC never know the optimial pixmap size, this would require some work to fix.

Other than that, I would still like to see some of the ::get* going away for API consistency with the LRC/Qt nameing guidelines (see the LRC wiki (it may still be in sflphone wiki))

Patch-set: 7
Label: Code-Review=0