translations: compile and install for lrc
This patch adds support via Cmake for compiling .ts translation files into
the .qm files which are used at runtime. These files are installed into:
<install prefix>/share/libringclient/translations dir, following the qt lib
convention. The .ts files are expected to be in the translations/ dir,
following the 'lrc_<language code>.ts' naming convention.

An initial French language translation file is included in the patch as a
proof of concept.

Issue: #78928
Change-Id: If9c1e8520ccee58ed8ce573bd2703e029fcc7b37