Update patch set 2
Patch Set 2:

In the KDE client (and probably all clients should do that), it is possible to query the errors of an account and the registration logs. This is implemented by many SIP client to check if there was a downtime when connected to a SIP account. Downtime can induce "invisible" missed calls. It is next to the account status string in the KDE client.

The string can also be displayed when registration fail, there is an API for that and it is used. The whole AccountStatusModel is there for that. Unless you found a way to translate *all* of those strings in pjsip itself, please revert the commit. And yes, some string *do* (like "no such file or directory") get translated by the system on some platforms, but not all (AFAIK: Windows).

Patch-set: 2