osx: fix crash on quit
On exit, ContactMethod are destroyed by their QObject parent,
PhoneDirectoryModel instance.

Implemeting a unique_ptr on top of it for m_pDialNumber caused a double free.

First the PhoneDirectoryModel is destroyed, and so all of its children.
Then Calls are destroyed by the CallModel, and the unique_ptr mechanism is
called, but on dealloced ContactMethod.

Issue: #79948
Change-Id: I68fb966d080f4c993ca83c68703d5dfb470f5ad9