Update patch set 6
Patch Set 6: Code-Review-1

While I am fine with the intend of the change (well, I still think the 24h based one is more useful, but whatever), I would argue against the implementation. The code used to be like that, but profiling showed that this caused a bottleneck when starting the application. This is due to the fact that the Julian calendar has a funky history and a lot of corner cases. So using QDate for this cause a noticable slowdown of the application (I don't have the Valgrind of this anymore, sorry). This was later partially mitigated by caching the result, greatly reducing the number of calls when performing the initial sorting. However, this also introduced the bug where items don't migrate from categories to categories. So that caching should eventually be eliminated again.

It is why the modulo/time_t based code should still be used. Unix time is free from the Julian/Gregrorian legacy, ignore leap seconds and calls older than 1902 is let say "unlikely". I know this sound weird, but you opened a little pandora box here. It should not be too hard to get the "seconds since midnight" using some current time arthimetic. Other than that, non-inlined functions call should be avoided and the other usual critical code path tricks should stay there. A before/after profiling with the caching turned off should also be made.

Patch-set: 6
Label: Code-Review=-1