Update patch set 1
Patch Set 1:

Look good to me. However, I wonder if it could be "upstreamed" in Person/Call/CM/IM/etc classes. That was originally the point of roleData, but they never had a common set of roles (Call roles are used as the "common" set, but it is more a hack than something designed).

Maybe that enum class could be part of the global namespace, then you could add Qt::DisplayRole and EditRole "Clones" with the same value (aka enum class { ULTIMATE_DIAPLY_ROLE = Qt::DisplayRole, MY_OTHER_ROLE = Qt::UserRole+1, MY_NEXT_ROLE};).

Then add a new version of ::dataRole for each of those class that take an enum class parameter. This should not cause ambiguity as enum class and not implicitly converted to integers, so the existing ::roleData could remain unchanged.

It should not be too long to implement (just cut and paste of this code into the respective methods) and would put the logic "where it belong". What do you think?

Patch-set: 1
Label: Code-Review=0