Update patch set 2
Patch Set 2:

After looking at it, there is a bunch of very common ones that are used by most models. Those are mostly the ones that spawn Call and CM. They are the ones used in the PR and maybe some more.

RecentModel
HistoryModel
PersonModel
CallModel
BookmarkModel

mostly use this set in one way or another. I think this should be the initial ones shared

Patch-set: 2