Revert "account: fix for hostname settings for Ring Account"
This reverts commit e1e66fef5ec990059a633ed8cfa3202c5e760e58.

The Ring account hostname is a CSV (; separated) structured field.
It can be accessed (and parsed) by the bootstrap model.

It is understandable not to want to display the full bootstrap model.
After all, it take a loot of space and provide little value. If a
client wish to display the first line only, then it should be advised
to get it from the bootstrap model rather than using the hostname
field directly.

Change-Id: I3a35a8d164b107237e79c0eff76db212c8e82fa5