roles: add a global LRC item data role
Many LRC models expose the same types of data, but do so using their
own item data roles. For example, Call::Role::Object exposes the item's
object pointer in some models, but in the PersonModel it is
Person::Role::Object. Since their enum values are not guaranteed to be
the same, and indeed they are not usually the same, this often requires
additional logic in the clients of LRC to determine which model is
being used.

This patch adds a Ring::Roles enum to be used by LRC models when
exposing the same types of data as other models.

Issue: #81198
Change-Id: I506c716c12f5b1bb0707e37c6dfa709b4fe03864