Update patch set 1
Patch Set 1: Code-Review+2

I think a macro would make it nicer and dropping the numbers altogether could be a possibility (while mildly breaking the numeric constants anyway, why not also fix them. The only restriction it would add (and the reason why I used explicit numbers in the first place) is that it wont be possible to move them around anymore without silently breaking the ABI. Let say Call::Role::Object move from 100 to 200, the ABI will technically be the same, but a client compiled with the old A*P*I will be able to run the new A*B*I without crashing, but will have the wrong runtime QVariant from the ::data() method and will then probably crash. On the other hand, we don't currently ship a .so, so the numbering don't have to be frozen. It's a choice. Either way I am fine with it.

Patch-set: 1
Label: Code-Review=+2