recentmodel: override data() method of PeopleProxy
The PeopleProxy filters out the the first call under a contact in the case
that it is the only call. In this case we want users of the PeopleProxy to
still be able to get the status of this call so that, for example, the
front-end could display it.

Issue: #81004
Change-Id: I6c067ffe424911eb0a69531f988f8aa14e9e3430