recentmodel: dataChanged CM and Person changed
Connect to the changed signal of ContactMethod and Person objects
when creating their nodes so that dataChanged can be emitted by
the RecentModel when these objects are updated.

This adds getIndex(Person) and getIndex(ContactMethod) methods to
the RecentModel to facilitate this change and to complement the
getIndex(Call) method.

Issue: #81198
Change-Id: I73bad37dd78fc895c2a390c1ec6ded25cc8ce9e3