Update patch set 2
Patch Set 2:

> connect the signal to the node itself and have up "upstream" the signal after that?

If I understand correctly, to do that I would have to turn the RecentViewNode into a QObject, so that it can have a slot to connect to the changed() signal and then in turn emit a signal which the RecentModel would connect to, no? Is the overhead of turning it into a QObject worth it to avoid the lambda mapping?

Patch-set: 2
Label: Code-Review=0