recentmodel: refactor call handling
This patch simplifies the code path. Now all changes to the Call nodes
are triggered by connecting to the changed() signal of the Call object
instead of calling slotUpdate() from multiple places and removing the
call nodes by connecting to the lifeCycleStateChanged signal.

When a new Call is first added, we connect to its changed() signal and
then either add it into the RecentModel or the callbucket. When a
changed() signal is emitted on a Call, we first check if it needs to be
removed and then if it needs to be moved from the callbucket to the
RecentModel.

This coincidenatlly fixes a bug in which calls which were OVER were
not removed from the callbucket.

Issue: #81198
Change-Id: I811eedcdb0c25fa8bc3bb86db2ff7fb3733e8ce8