recentmodel: refactor call handling
This patch simplifies the code path and removes the "callbucket".

Now all changes to the Call nodes are triggered by connecting to the
changed() signal of the Call object instead of calling slotUpdate()
from multiple places and removing the call nodes by connecting to
the lifeCycleStateChanged signal.

When a new Call is first added, we connect to its changed() signal and
then add it into the RecentModel if it is already associated with a CM.
When a changed() signal is emitted on a Call, we check if it needs to
be removed or else added to the RecentModel.

This coincidenatlly fixes a bug in which calls which were OVER were
not removed from the callbucket and then added to the RecentModel.

Issue: #81198
Change-Id: I811eedcdb0c25fa8bc3bb86db2ff7fb3733e8ce8