recentmodel: update indeces when call is removed
They must be updated because the m_Index refers to the row
of the child in the model and is used to generate the
corresponding QModelIndex. Unless the removed call is the
last child, the m_Index of the other calls may be wrong
and they will point to the wrong QModelIndex.

Issue: #81454
Change-Id: I873e11f509eb9323cf153c84bcf9458595b17e74