source model: Bind a sourcemodel to a Call
The video source information is contained into the sourcemodel class.
This class is currently a singleton, so there is only one source
information. But the deamon handle sources management correctly: if you
switch from a call to another and then switch back to the first the
source will be the one selected by the user, but the client has lost
this information.

We need multiples instances of the source model class, one related with each
call to keep up with the daemon.

This patch removes the singleton pattern from sourcemodel, creates
instances for each calls and add a utility static method
SourceModelFromSelectedCall() to keep the same baheviour for the clients
as when they where invoking the singleton.

Change-Id: Ie531c5e8ae432f189f434d0ca3a963f1fb0a11ef
Tuleap: #24