fix singleton implementation
This patch re-implements all Singletons using Construct On First Use Idiom
in a re-entrant way.

This also changes the API of Singleton instance() method by returning
a reference than a pointer. This specify that API is not supposed to
return a null pointer (or here a null reference).

Some parts have been modified as this Singleton implementation has
showed some not-well documented issues (like class loop-dependencies
on Account and AccountModel).

Tuleap: #44
Change-Id: I0061c7c0302f24e13d49ecfd125def84e24660d2