remove all dangerous globals
We may have "static initialization order fiasco" due to numerous globals,
specially with Singletons implementation based on globals.

This patch re-implements all globals using Construct On First Use Idiom.
This also changes the API of Singleton instance() method by returning
a reference than a pointer. This specify that API is not supposed to
return a null pointer (or here a null reference).

Tuleap: #44
Change-Id: I0061c7c0302f24e13d49ecfd125def84e24660d2