Update patch set 22
Patch Set 22:

I understand that you re interested in it but some things are wrongly placed or written for testing purpose (because we encountered a lot of bugs on the road...) So don't waste your time checking code quality for now. Save it for the release candidate, i'll be all ears.

We aim to make conference and transfer a user friendly experience and this patchset aim to implement what we need to make the flow logical.

Patch-set: 22