recentmodel: fix bad move handling
- Note that if sourceParent and destinationParent are the same,
you must ensure that the destinationChild is not within the range
of sourceFirst and sourceLast + 1. You must also ensure that you do
not attempt to move a row to one of its own children or ancestors.
This method returns false if either condition is true, in which
case you should abort your move operation.

Change-Id: I6af4dcbbaf648343d1f7a5d31610d0b2ea0c9713
Tuleap: #48