Update patch set 10
Patch Set 10: Code-Review-1

There is a lot of good in this. Some of the most recent singleton were written that way and it removed many issues you solve with this. Also, while improbably, some of the re-entry issues this solve did indeed bit me in the past.

However, I am not a fan of returning references

First, this break the API for no good reasons and make it inconsistent as exported methods are forbidden to return references due to this KDE code policy:

https://techbase.kde.org/Policies/Library_Code_Policy#Const_References

While the issue raised by the policy aren't really relevant in this case, I would prefer not to turn it off. This policy was/is (for me) enforced by static analysis. Merging this patch would require me to turn the whole policy off.

Otherwise, this is a good change

Patch-set: 10
Label: Code-Review=-1