recentmodel: prevent use after free
This removes code which tried to move child calls of a CM to
the Person with which it is now associated with. This code did
not work properly as the proper signals were not emitted and it
also caused a use after free as the children were not removed
from the CM which is then deleted and deletes all its children.

Right now, in most cases the calls will be moved to the new Person
correctly during slotChanged(); however a more comprehensive
solution requires some more work.

Change-Id: Ia0542ace9e8cc7df347dac8b803d8a7283fec753
Tuleap: #73