*WIP* recentmodel: sort using proxy model
This works, however the QSortFilterProxyModel emits
layoutChanged instead of rowsMoved most of the time when
sorting dynamically, which doesn't work well with the
GNOME and OSX clients.

It is simpler to sort using the existing functionality
in Qt. It can also be easily extended later to sort
(and filter) using other criteria as well.

Change-Id: Ic7981ccbd801969439853db284f8da72f1f073ec
Tuleap: #84