Update patch set 2
Patch Set 2: Code-Review+2

This class is one of the last without a selection model. This is due to the fact that is was written for Qt4 that didn't have selectionModel in QtCore. The code was never ported to Qt5.

But as this have nothing to do with the change itself, +2 anyway

Patch-set: 2
Reviewer: Gerrit User 11 <11@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=+2