Update patch set 1
Patch Set 1:

Open question:

If the person has a "rich" history with audio/video recording and mb of chat text, don't you want to see that even if the person cannot be reached?

One scenario is where a SIP account is valid at work, but will fail to connect from home (assuming the users isn't connected to a VPN in a modern BYOD scenario). Maybe this could be a role and differenced by the UX painter rather than just disabling/hiding the item?

Patch-set: 1
Label: Code-Review=0