Update patch set 2
Patch Set 2:

I thought that isReachable will only check the availability of an account, not whether or not it is connected... but you're right that if there is a history with the contact, it should probably still be shown. And going further, we should probably distinguish contacts which we can't reach because the account is dissabled or not registered. But, not in this patch...

Patch-set: 2
Label: Code-Review=0