callmodel: generic insert of non-existing calls from daemon
This patch adds foreign calls seen when their state is changed.
These calls come from daemon when this one is the creator of them.
In this case LRC is not aknowledged of their existance.
So this patch forces LRC to handle them when the event
"CallStateChanged" is triggered by the daemon, than just ignore them,
as was done before.

This patch also fixes the unattended transfer (aka "blind transfer")

Change-Id: I8bd502c77363e854c7b466ddbf357b29b2fb9f46
Tuleap: #93