Update patch set 5
Patch Set 5:

(11 comments)

Here is the review. Don't be scared by it, this thing is getting closer and the earlier issues with the previous patches have mostly been fixed (that's good).

Maybe take a closer look at both LRC and Qt coding/naming guidelines to avoid the my little nitpicking comments about naming. Just keep in mind, the point of that is keeping the API and code consistent across the library.

Here is the API doc for the selectionmodel thing http://doc.qt.io/qt-5/qitemselectionmodel.html

Patch-set: 5