video: fix bad class construction order
Classes in Video::ConfigurationProxy have some dependencies between them.
But these dependencies are not forced by construction: the code
uses raw pointer than using getters; these lasts construct needed instances
if not yet done (Singletons).

A problem of this fact is if, for example, deviceModel() is called after
deviceSelectionModel(), the latter is not correctly initialized as it
depends of construction of the former.
This is the bug described by tracker Tuleap#17.

The patch fixes this fact by calling getters when the corresponding instance
is requiered than using direct raw pointer.

A big thanks to Ales Katona <almindor@gmail.com> who permits the discovery
of this bug.

Change-Id: I60d8880676cef1cf3c8b56fefbf503ac32025eb6
Tuleap: #17