im: finish LRC out of call API
- migrate to payload API to allow for multipart text message as per
daemon API change
- connect incoming message signal in the phonedirectorymodel
and create a signal for client to listen to
- save out of call text message by using the already existant
instantMessagingModel
- always load text recording when we ask for it

In the context where text messages can now be transmitted outside
of a call, the TextRecording object must be easier to manipulate and
create outside the context of a call.
this patch does remove a limitation regarding such creation.

Tuleap: #152
Change-Id: Iefaac4e23f7cb3f2530fb2937b17e32fa8616bb4