call: fix invalid state on init of existing call
- Remove DaemonStateInit class: duplicates DRing::Call::StateEvent
- Add missing DRing::Call::StateEvent in startStateFromDaemonCallState
  => This fix an assert caused by existing call being created
  with failure state instead of connecting.

- add support for new state introduced in daemon

This commit also fix 3 serious bugs introduced in July and November

The first one is related to the addition of the INACTIVE daemon state.
This states wasn't handled in the LRC state machine and caused many
invisible (but in the logs) errors

The second and third ones where introduced when it was allowed to
create a call before the "incomingCall" signal. As the sagnal was
emited later, the call was created twice and one copy was sent
to the history (due to bug #1) and the second added to the CallModel.

This, in turn, caused the history call (as introduced first) to receive
many signals, causing yet more subtle issues. Those issues were made
visible by the addition of the OVER signal, as it was handled by the
wrong duplicate call.

Tuleap: #169
Change-Id: If84452169848172b9097422b6988ec7a30a7480d