model: Handle Qt::DecorationRole in roledata()
When clients require simple widgets that filter "raw"
model data, such as selecting a contact when receiving
a call from an unknown contact method, it is overkill
for them to have to implement a view delegate to display
the picture.

This commit add DecorationRole to all models based on
Ring::Role. The pixmap manipulation interface need to
implement the new functions to make use of this.

The interface is responsible to support resolution
independance.

Tuleap #170

Change-Id: Ibad44805485b71195805eaca82f08a4a56864f50