Update patch set 6
Patch Set 6:

As promised, I tested, there is a good news and a bad news. The good news is that your patch is good!

However, the bad news is that the previous patch that added INACTIVE was not, so your patch break the state machine. The call is actually currently created OVER and is added to history before being CURRENT. That, in turn, mess things up considerably... I will do a quick fix (its 3am), but it will require even more test as the change will be more substantial.

Patch-set: 6
Label: Code-Review=0