textrecording: Make it easier to know a new message has arrived
With the new offcall texting, it is necessary to know that a text
has arrived. For call, this come from CallModel, but there is no
alternative for "pure" Media::TextRecording.

This commit add such alternative to Media::RecordingModel, the
recording manager (currently mostly unused)

THIS COMMIT NEED SERIOUS REBASING, I KNOW **DNM AS IS**

Change-Id: I3ff426e8675380f424ac24e2bee2d7bcd2fb992a
Tuleap: #190