Update patch set 2
Patch Set 2:

If you want offcall messages, there was no way (beside maybe the recentmodel) to know about it. The "old" way to know a message has arrived was to connect to CallModel::mediaAdded and then to connect to that media.

This commit add an alternate way to get all TextRecording that changed. With or without a call. The API is not as good as it could be, because there is no object for that and the whole "textmessage" name is not really representative of that this class can do. Beside that, it work fine and give the information when a client needs it.

Patch-set: 2