numbercompletion: Show IP2IP when relevant
The rational for not showing IP2IP was that any account could
do IP2IP calls as the daemon detect it internally. However, it is
deeply confusing when you really try to call an IP.

Therfor, this commit restore showing IP2IP when the URI::ProtocolHint
match an IP address.

Change-Id: Ic105b3b230407a6e74a6923a6c971074cebda829