credentials: Remove the union
It was an experiment, it failed. The number of cache fault
avoided by it is negligeable. It make debugging a pain and
prevent the use of "new" and "delete". Then, there is still
a crash in the freeing sequence.

I could have fixed it, but would rather fix the cause: unions

Issue: #80868
Change-Id: I107102f05d151055617fd4256e07ed5aaa196b2e