ringtone: Implement missing bits
When porting all objects to the Collection system, the RingtoneModel
gained the ability to keep a global list of RingTones, add folders
and so on. However, this was never finished due to time constraints.

This commit complete the basic features. It is now good enough to be
used by the clients.

Tuleap: #204
Change-Id: I36a14b811b36646aa62f87aeccdbab130f416f43