Update patch set 2
Patch Set 2: Code-Review-1

Is this the patch that is supposed to fix Ring contact methods not showing up in the number completion model when searching for them using the name of the contact? Or was that another one? Just asking because I can't seem to reproduce the issue any more with or without this patch...

Also, if apply() can be called from multiple threads, couldn't getNumber still be called before the contact is fully loaded? Or am I missing somethign.

Patch-set: 2
Label: Code-Review=-1