Update patch set 2
Patch Set 2:

This patch does solve some auto-completion corner case.

As for getNumber, this is not an issue if it is called without a Person. The bug only happen when it is called with an incomplete Person as the UID is compared in the ::getNumber algoritm

Patch-set: 2
Reviewer: Gerrit User 11 <11@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=0