history: Fix crash in time calculation after Jan 1 or 6 months
This is a regression from:
e0ff769e1980f19486843da7ded91ab9f613a58e

This commit revert the changes and upgrade the older and better tested
code path. It also upgrade it to drop the nychthemeron algorithm for
a midnight aligned one. This simplified the logic as more modulo
could be used.

This code also add support for timezones. This is a new feature and
the issue was detected during (quite extensive) testing using mock
history calls. It still has a bug if the timezone have both standard
and daylight saving time. This commit don't try to fix it as there is
no solution for this beside living with it. It only affect calls made
between 23h and 1h. It will only cause call from "yesterday" to be
classified as 2 days ago. This isn't user visible enough to try to
track the date of each year daylight time saving switch.

I considered fixing the new code path, but there is more than 1
underflow and rely on external code. This new code is the best of
both world. It use the same (simple) logic as the new code path
while being as bullet proof as the old one. The rewrite may have
introduced new issues, but I don't think so and none of my fake
calls or any of my 2500 history calls from the last 2 years
caused issues.

Change-Id: I2ad4724f733b2722899bba376dc2e62135544625
Tuleap: #214