Update patch set 2
Patch Set 2:

> This code seems to reinvent the weel. Why not using system or library provided datetime computation methods ? E.g http://doc.qt.io/qt-5/qdatetime.html

It was, see:
https://github.com/Elv13/libringclient/blob/98313595d643caff790ed77234b928e852e51e62/src/historymodel.cpp#L443

From May 2013. The problem was, this code was/is called during sorting, and was the bottleneck. Some other aspect of the sorting have since been optimized, indirectly reducing the number of calls to this function. But even when I added sorting caching, I tried to "downgrade" to the old code path (from 2011) again, but it was still showing high in Valgrind CallGrind, so I kept the optimized version.

This code is O(1) and never avoid calling external functions for nothing. QDateTime functions are not inlined (they are .so symbols, they can't be) and the gredorian calendar have more corner cases than seconds since 1970, making it often slow (for correct implementation, most "recent" implementations ignore the leap seconds, skipped days, special leap years, changes in local daylight dates, timezone swaps and all other wierdness)

Patch-set: 2