Update patch set 3
Patch Set 3: Code-Review-1

I really don't think this is the way to go. This code was in production for a while and had issues, it will have the same issues again. The Moore law is not an excuse for re-introducing slow code.

You replaced a division with a for loop, you can't call that optimized. It is orders of magnitude slower and in a critical code path.

I also don't think it is simpler than the other one. Once you understand you can use modulo on timestamps to get days. seconds since midnight and elapsed days, it make sense.

Patch-set: 3
Label: Code-Review=-1