Update patch set 2
Patch Set 2:

> coudn't this only be called for display and simple comparaison used for sorting ?

No, the items need to be categorized before being sorted, then once categorized, they are sorted

> try this optimized version of the QDate code

Again, this was slow in 2013, this is still slow today, and you are using for loops, and creating 30 temporary QTime object on the stack for no reasons

My version create nothing on the stack beside the current timestamps. It it calls no function beside getting the system time and have no loops. I could make it a little more readable if it is your convern

Patch-set: 2