Update patch set 3
Patch Set 3:

It is optimized compared to the previous code using QDate that had performance issues :
https://github.com/Elv13/libringclient/blob/98313595d643caff790ed77234b928e852e51e62/src/historymodel.cpp#L443
The algorithm now treats most frequent (instead of less frequent) cases first.
It avoids entering loops for unrelated dates, and doesn't make a call to QDate::currentDate() for every date comparison. It also reduces the number of comparison made in loops.

Of course this optimization will still be slower than the handmade version but it handles all the calendar subtleties, and should be about 10x faster than the previous code with a large history.
As an example, classifying a call from last year will result in 
1  call  to QDate::currentDate() and 2  date comparisons, instead of 
38 calls to QDate::currentDate() and 38 date comparisons.

I think the goal here should be to have a code that works fine in every scenario including year changes, leap days (we have one soon in February), weird times, weird timezones etc., and with good enough speed.

Patch-set: 3
Label: Code-Review=0