Update patch set 1
Patch Set 1: Code-Review+1

Thanks.

Technically, this should be ported to the state machine pattern. There is a race condition where the device can change while the user is configuring them. The probability is this are quite low, but not void. The chances that the user plug an headset while configuring the client is actually higher than "background noise". 

But then, this patch is good enough already for 99% of the case and the state machine implementation have a steep learning curve, so it is probably not worth the time. See the credentialmodel or the codecmodel to see 2 examples of what would have to be done.

Patch-set: 1
Reviewer: Gerrit User 11 <11@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=+1