Update patch set 1
Patch Set 1: Code-Review-1

Placeholders are not "terrible" per se, they solve a specific problem.

When you concurrently load linked data from multiple sources, some will arrive first. You can use complex logic to handle this, or you can use placeholders. A placeholder will be upgraded to the real object when it arrive (if ever). This is why contact work in LRC. Without it, there would be no way to make it work beside saving everything in sqlite, including contacts (duplicate the system store: bad idea).

The issue here is not the placeholder. The issue is that contactmethods should check if the account is a placeholder before using it. Placeholders should also not be deleted in any case. The point is having a valid pointer for objects to hang to in case the real one arrive.

Remember, LRC create an object database in the background, the links between objects are very important. If you prevent those links from being created (in this case, the profile association), everything will start to break. The message is "suppress dangerous placeholder", but I would argue that when used correctly, they are not dangerous. The Person:: implementation is safer than this one, but they both serve a purpose. This one is not used globally, TextRecordings should not use placeholders. That would create ghost CM in the PhoneDirectoryModel. But for profiles, it does what it does.

By the way, if you delete it, also remove the ::merge method of Account.

-1 for breaking a working feature without a fix in the pipeline, the KDE client use profiles.

Patch-set: 1
Label: Code-Review=-1