Update patch set 1
Patch Set 1:

It's actually because it's not safe AT ALL that i say it's a terrible idea. The placeholder concept in itself is not dangerous. BUT i still think it was not a good idea to load all that data concurrently, and after all the problem we've had with the contact placeholder i think we all agree that loading all that data step by step would have spare a whole level of complexity that are time consuming for our day to day operations. 

You're the only one using profile at the moment so feel free to propose something to replace that. We don't really have much time, we're all in a "stress test/fix bugs/stabilization" phase before FOSDEM.

You're right i did see the merge method but forgot to delete it.

Patch-set: 1
Reviewer: Gerrit User 32 <32@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=0